chore(procmgr): drop redundant spawn identity wrapper module
Remove spawn/identity.rs and colocate spawn-user assertions with the
platform intended_spawn_user helpers that already own that behavior.
